# Add multiprocess support to `TrajectorySplitters`