Closes #461 

- There are some minor internal API changes to facilitate this but nothing on the user-facing end.
- The tests are taken from the single-processing version and expanded to use multiprocessing. I hope that is okay.
- Also, renamed a test method inside `test_trajectory_splitter.py` to follow proper naming convention.

Let me know what you think about it, @anitagraser