@anitagraser 2025-05-07
For future reference, here are some performance comparisons

Mode | min_length | Early Skipping | n_processes | 0.01 million points (s) | 0.1 million points (s) | 0.5 million points (s)
-- | -- | -- | -- | -- | -- | --
Value Change Splitter | 1000 | No | 1 | 2.632380 | 24.794708 | 132.504514
Value Change Splitter | 1000 | No | 16 | 5.792837 | 18.799301 | 77.033445
Value Change Splitter | 1000 | Yes | 1 | 2.021787 | 17.722349 | 97.096850
Value Change Splitter | 1000 | Yes | 16 | 5.602232 | 17.298913 | 69.297309
Value Change Splitter | No | No | 1 | 1.989890 | 21.177990 | 89.591676
Value Change Splitter | No | No | 16 | 5.037065 | 7.698795 | 21.954373

(Early skipping was introduced in https://github.com/movingpandas/movingpandas/pull/458)


![image](https://github.com/user-attachments/assets/191d76bb-06f1-41a6-a8a5-2073cc38a36c)




```


### Lazy init + Value Change Splitter (n_processes=1)

#### Without min_length

--- 0.01 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  1.989890s

--- 0.1 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  21.177990s

--- 0.5 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  89.591676s

#### with min_length 1000 

--- 0.01 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  2.632380s

--- 0.1 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  24.794708s

--- 0.5 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  132.504514s

#### with min_length 1000  and early skipping

--- 0.01 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  2.021787s

--- 0.1 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  17.722349s

--- 0.5 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  97.096850s




### Lazy init + Value Change Splitter (n_processes=16)


#### Without min_length

--- 0.01 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  5.037065s

--- 0.1 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  7.698795s

--- 0.5 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  21.954373s

#### with min_length 1000 


--- 0.01 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  5.792837s

--- 0.1 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  18.799301s

--- 0.5 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  77.033445s


#### with min_length 1000 and early skipping

--- 0.01 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  5.602232s

--- 0.1 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  17.298913s

--- 0.5 million points ---
TrajectoryCollection with 16 trajectories
Trajectory init + ValueChangeSplitter:  69.297309s


```